7주차- 경과시간 및 생성일 구하는 컴포넌트 분리
