+++ src/components/Card.js
@@ -1,34 +1,7 @@
 import noneCardImg from '../assets/none-card-img.png';
+import ElapsedTime from '../utils/ElapsedTime';
+import FormatDate from '../utils/FormatDate';
 import './Card.css';
-
-const formatDate = (value) => {
-  const date = new Date(value);
-  return `${date.getFullYear()}. ${date.getMonth() + 1}. ${date.getDate()}`;
-};
-
-const elapsedTime = (createdAt) => {
-  const milliSeconds = new Date() - new Date(createdAt);
-  const seconds = milliSeconds / 1000;
-
-  const diffNum = (value) => Math.floor(value);
-
-  if (seconds < 120) return `1 minute ago`;
-
-  const minutes = seconds / 60;
-  if (minutes < 60) return `${diffNum(minutes)} minutes ago`;
-
-  const hours = minutes / 60;
-  if (hours < 24) return `${diffNum(hours)} hour ago`;
-
-  const days = hours / 24;
-  if (days < 30) return `${diffNum(days)} day ago`;
-
-  const months = days / 30;
-  if (months < 12) return `${diffNum(months)} months ago`;
-
-  const years = days / 365;
-  return `${diffNum(years)} years ago`;
-};
 
 const Card = ({ link }) => {
   const { createdAt, description, imageSource } = link;
@@ -48,9 +21,9 @@ const Card = ({ link }) => {
           )}
         </div>
         <div className="card_contents_area">
-          <p className="elapsed_time">{elapsedTime(createdAt)}</p>
+          <p className="elapsed_time">{ElapsedTime(createdAt)}</p>
           <p className="description_text">{description}</p>
-          <p className="created_at">{formatDate(createdAt)}</p>
+          <p className="created_at">{FormatDate(createdAt)}</p>
         </div>
       </a>
     </div>
